requirements.txt
bme280==0.7
paho_mqtt==2.0.0
psutil==5.9.8
smbus2==0.4.3
